fixed MSVC 2008 compile errors and warnings
git-svn-id: file://localhost/home/tolik/svnrepo/trunk@2983 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ opencv/3rdparty/lapack/slamch.c
@@ -78,7 +78,7 @@ slamch_(char *cmach) {
         /* Use SMALL plus a bit, to avoid the possibility of rounding causing overflow
          when computing  1/sfmin. */
         sfmin = FLT_MIN;
-        small = 2. / FLT_MAX;
+        small = (float)(2. / FLT_MAX);
         if (small <= sfmin) small = sfmin * (1 + eps);
         return small;
     } else if ('U' == ch || 'u' == ch) {
